elliptic-curve: SEC1 private key support
Adds the following methods to `SecretKey`:

- `from_sec1_der`
- `from_sec1_pem` (requires `pem` feature)
- `to_sec1_der`
- `to_sec1_pem` (requires `pem` feature)

These implement the SEC1 (RFC5915) ASN.1 DER encoding for elliptic curve
private keys.

These methods are extracted from and now consumed by the PKCS#8
implementation, as PKCS#8 is just a wrapper for this format when used
with elliptic curve keys.

It seems like it might be a good idea to eventually extract this into a
`sec1` crate as a followup, providing a similar role to the `pkcs1`
crate. This would allow a setup that works much more like the `rsa`
crate does now, where impling `sec1` traits provides a blanket impl of
PKCS#8 support.
